pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nooscope"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "watchdog",
    "pyyaml",
    "python-frontmatter",
    "mcp",
    "httpx",
    "numpy",
]

[project.optional-dependencies]
apple = ["pyobjc-framework-NaturalLanguage"]
mlx = ["mlx-lm"]
openai = ["openai"]

[project.scripts]
nooscope = "nooscope.cli:main"
